Person updating done

+++ PhoneBook/Form1.cs
@@ -14,6 +14,7 @@ namespace PhoneBook
     public partial class MainForm : Form
     {
         private string currentPhotoPath;
+        private Person CurrentPerson;
 
         public MainForm()
         {
@@ -79,16 +80,16 @@ namespace PhoneBook
             string _SurName = personItem.Substring(index+1, personItem.Length-index-1);
             string _GroupName = this.groupsComboBox.SelectedItem.ToString();
 
-            Person anyPerson = new Person(_GroupName, _Name, _MiddleName, _SurName, true);
-            this.nameTextBox.Text = anyPerson.Name;
-            this.surNameTextBox.Text = anyPerson.Surname;
-            this.MidNameTextBox.Text = anyPerson.MiddleName;
-            this.mobilePhoneTextBox.Text = anyPerson.MobilePhone;
-            this.homePhoneTextBox.Text = anyPerson.HomePhone;
-            this.workPhoneTextBox.Text = anyPerson.WorkPhone;
-            this.emailTextBox.Text = anyPerson.Email;
-            this.addrTextBox.Text = anyPerson.Addres;
-            this.currentPhotoPath = anyPerson.PhotoPath;
+            CurrentPerson = new Person(_GroupName, _Name, _MiddleName, _SurName, true);
+            this.nameTextBox.Text = CurrentPerson.Name;
+            this.surNameTextBox.Text = CurrentPerson.Surname;
+            this.MidNameTextBox.Text = CurrentPerson.MiddleName;
+            this.mobilePhoneTextBox.Text = CurrentPerson.MobilePhone;
+            this.homePhoneTextBox.Text = CurrentPerson.HomePhone;
+            this.workPhoneTextBox.Text = CurrentPerson.WorkPhone;
+            this.emailTextBox.Text = CurrentPerson.Email;
+            this.addrTextBox.Text = CurrentPerson.Addres;
+            this.currentPhotoPath = CurrentPerson.PhotoPath;
             if (this.currentPhotoPath == null)
                 this.currentPhotoPath = CurrentDirrectoryReturner.getDirrectory() + "\\data\\noPhotoAvailable.jpg";
             var picStream = File.OpenRead(this.currentPhotoPath);
@@ -98,19 +99,77 @@ namespace PhoneBook
 
         private void saveButton_Click(object sender, EventArgs e)
         {
-            string _GroupName = this.groupsComboBox.SelectedItem.ToString();
-            string _Name = this.nameTextBox.Text;
-            string _MiddleName = this.MidNameTextBox.Text;
-            string _SurName = this.surNameTextBox.Text;
-            Person anyPerson = new Person(_GroupName, _Name, _MiddleName, _SurName, true);
-            Dictionary<String, String> UpdatedValues;
-            //int res = anyPerson.update(String[] _UpdatingFiel, String[] _UpdatingValue, int _CountOfUpdating);
-            String newPhotoPath = CurrentDirrectoryReturner.getDirrectory() + "\\data\\Photos\\"+
-                anyPerson.Groupname+"_"+ anyPerson.Surname+"_"+anyPerson.Name+"_"+anyPerson.MiddleName +".jpg";
-            if (!anyPerson.PhotoPath.Equals(newPhotoPath)) {
-                File.Copy(currentPhotoPath, newPhotoPath);
+            Dictionary<String, String> UpdatedValues = new Dictionary<String, String>();
+            if (!this.groupsComboBox.Text.Equals(this.CurrentPerson.Groupname))
+                UpdatedValues.Add("groupName", this.groupsComboBox.Text);
+            if (!this.nameTextBox.Text.Equals(this.CurrentPerson.Name))
+                UpdatedValues.Add("name", this.nameTextBox.Text);
+            if (!this.surNameTextBox.Text.Equals(this.CurrentPerson.Surname))
+                UpdatedValues.Add("surname", this.surNameTextBox.Text);
+            if (!this.MidNameTextBox.Text.Equals(this.CurrentPerson.MiddleName))
+                UpdatedValues.Add("middlename", this.MidNameTextBox.Text);
+            if (!this.mobilePhoneTextBox.Text.Equals(this.CurrentPerson.MobilePhone))
+                UpdatedValues.Add("mobile_phone", this.mobilePhoneTextBox.Text);
+            if (!this.workPhoneTextBox.Text.Equals(this.CurrentPerson.WorkPhone))
+                UpdatedValues.Add("work_phone", this.workPhoneTextBox.Text);
+            if (!this.homePhoneTextBox.Text.Equals(this.CurrentPerson.HomePhone))
+                UpdatedValues.Add("home_phone", this.homePhoneTextBox.Text);
+            if (!this.emailTextBox.Text.Equals(this.CurrentPerson.Email))
+                UpdatedValues.Add("email", this.emailTextBox.Text);
+            if (!this.addrTextBox.Text.Equals(this.CurrentPerson.Addres))
+                UpdatedValues.Add("address", this.addrTextBox.Text);
+            if (!this.currentPhotoPath.Equals(this.CurrentPerson.PhotoPath) && !this.currentPhotoPath.Equals(CurrentDirrectoryReturner.getDirrectory() + "\\data\\noPhotoAvailable.jpg"))
+            {
+                try
+                {
+                    String _GroupName = this.groupsComboBox.Text;
+                    String _Name = this.nameTextBox.Text;
+                    String _MiddleName = this.MidNameTextBox.Text;
+                    String _SurName = this.surNameTextBox.Text;
+                    String _PhotoPath = CurrentDirrectoryReturner.getDirrectory() + "\\data\\Photos\\" +
+                    _GroupName + "_" + _SurName + "_" + _Name + "_" + _MiddleName + ".jpg";
+                    File.Copy(this.currentPhotoPath, _PhotoPath, true);
+                    UpdatedValues.Add("photoPath", _PhotoPath);
+                }finally {}
             }
-                       
+                
+            if (UpdatedValues.Count > 0)
+            {
+                int res = this.CurrentPerson.update(UpdatedValues);
+                if (res == 0)
+                {
+                    this.updateGroupsList();
+                    this.updatePersonsList(); 
+                    MessageBox.Show("Изменения сохранены", "Успешно!");
+                }
+                else
+                {
+                    MessageBox.Show("Ошибка сохранения", "Ошибка");
+                }
+            }
+            else
+            {
+                MessageBox.Show("Отображаемые данные актуальны. Нет изменений для внесения в базу данных", "Внимание");
+            }
+            /*
+            if (!this.currentPhotoPath.Equals(this.CurrentPerson.PhotoPath) && !this.currentPhotoPath.Equals(CurrentDirrectoryReturner.getDirrectory() + "\\data\\noPhotoAvailable.jpg"))
+            {
+                try
+                {
+                    String _GroupName = this.groupsComboBox.Text;
+                    String _Name = this.nameTextBox.Text;
+                    String _MiddleName = this.MidNameTextBox.Text;
+                    String _SurName = this.surNameTextBox.Text;
+                    String _PhotoPath = CurrentDirrectoryReturner.getDirrectory() + "\\data\\Photos\\" +
+                    _GroupName + "_" + _SurName + "_" + _Name + "_" + _MiddleName + ".jpg";
+                    File.Copy(this.currentPhotoPath, _PhotoPath, true);
+                    MessageBox.Show("Фото обновленно успешно", "Успешно");
+                }
+                catch (Exception exp)
+                {
+                    MessageBox.Show("Ошибка замены фотографии", "Ошибка");
+                }
+            }*/
         }
 
         private void addButton_Click(object sender, EventArgs e)
@@ -127,7 +186,7 @@ namespace PhoneBook
                 String _Email = this.emailTextBox.Text;
                 String _Addres = this.addrTextBox.Text;
                 String _PhotoPath = CurrentDirrectoryReturner.getDirrectory() + "\\data\\Photos\\" +
-                _GroupName + "_" + _SurName + "_" + _Name + "_" + _MiddleName + ".jpg"; ;
+                _GroupName + "_" + _SurName + "_" + _Name + "_" + _MiddleName + ".jpg"; 
                 Person newPerson = new Person(_GroupName, _Name, _MiddleName, _SurName, _MobilePhone,
                    _HomePhone, _WorkPhone, _Email, _Addres, _PhotoPath);
                 int res = newPerson.save();
@@ -168,22 +227,29 @@ namespace PhoneBook
             string _MiddleName = this.MidNameTextBox.Text;
             string _SurName = this.surNameTextBox.Text;
             Person anyPerson = new Person(_GroupName, _Name, _MiddleName, _SurName, true);
-            try
+            if (this.CurrentPerson != null)
             {
-                int res = anyPerson.delete();
-                if (res == 0){
-                    this.clearGUI();
-                    this.updatePersonsList();
-                    if (anyPerson.PhotoPath != null)
-                        File.Delete(anyPerson.PhotoPath);
-                    MessageBox.Show("Успешное удаление!", "Успешно!");
+                try
+                {
+                    int res = anyPerson.delete();
+                    if (res == 0)
+                    {
+                        this.clearGUI();
+                        this.updatePersonsList();
+                        if (anyPerson.PhotoPath != null)
+                            File.Delete(anyPerson.PhotoPath);
+                        MessageBox.Show("Успешное удаление!", "Успешно!");
+                        this.CurrentPerson = null;
+                    }
+                    else
+                    {
+                        MessageBox.Show("Ошибка! Удаление не было завершено! Попробуйте повторить попытку!", "Ошибка!");
+                    }
                 }
-                else{
+                catch (Exception exp)
+                {
                     MessageBox.Show("Ошибка! Удаление не было завершено! Попробуйте повторить попытку!", "Ошибка!");
                 }
-            }
-            catch (Exception exp){
-                MessageBox.Show("Ошибка! Удаление не было завершено! Попробуйте повторить попытку!", "Ошибка!");
             }
         }
     }
